CSS and new routes

+++ application/djangoapp/forms.py
@@ -1,4 +1,4 @@
 from django import forms
 
 class NameForm(forms.Form):
-    amount = forms.IntegerField(label='Montant')
+    amount = forms.DecimalField(label='Montant')
